pyproject.toml
[project]
name = "kafka-nats-benchmark"
version = "0.1.0"
description = "Kafka vs NATS JetStream restricted-hardware benchmark suite"
requires-python = ">=3.11"
dependencies = [
    "aiokafka>=0.11.0",
    "confluent-kafka>=2.6.0",
    "nats-py>=2.9.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.9.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-packages = true
packages = ["src"]
